flask_backend/requirements.txt
flask
paho-mqtt
pillow
pytesseract
flask-socketio
eventlet
python-dotenv